update: graphFunctions.tsから関数を移動＆不要ファイルを削除
